valgrind: adjust std_list expected output
The run-ptest script uses a custom --yocto-ptest option
to change the summary statistics that valgrind issues
to a format that matches what ptest-runner expects.

Only this std_list test which checks that strcpy() works
properly on Ubuntu-12.04 and later is affected by this change.
It's normal output when called with --yocto-ptest is a blank
line so adjust the expected output.

Upstream-Status: Inappropriate [Yocto specific]

Signed-off-by: Randy MacLeod <Randy.MacLeod@windriver.com>